Merge commit '6d494f9c88084a10e9ba835508b58a139fc74bd3'

* commit '6d494f9c88084a10e9ba835508b58a139fc74bd3':
  cdma sms encoding tweaks and logging / debugging annotations.

+++ telephony/java/com/android/internal/telephony/cdma/sms/UserData.java
@@ -154,7 +154,7 @@ public class UserData {
         builder.append("{ msgEncoding=" + (msgEncodingSet ? msgEncoding : "unset"));
         builder.append(", msgType=" + msgType);
         builder.append(", paddingBits=" + paddingBits);
-        builder.append(", numFields=" + (int)numFields);
+        builder.append(", numFields=" + numFields);
         builder.append(", userDataHeader=" + userDataHeader);
         builder.append(", payload='" + HexDump.toHexString(payload) + "'");
         builder.append(", payloadStr='" + payloadStr + "'");
